requirements.txt
pathway
openai
pandas
python-dotenv
google-generativeai
